# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Zymo API
  version: "1"
  description: Self-hostable fermentation tracking API. All unauthorized access returns 404, not 403, to avoid leaking resource existence.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: This instance

# All operations require auth by default. Public endpoints opt out with `security: []`.
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: zymo_session
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    CursorParam:
      name: cursor
      in: query
      description: |
        Opaque pagination cursor returned as `next_cursor` from a prior page.
        Omit (or pass empty) to fetch the first page. Format is intentionally
        not part of the API contract — clients should treat the value as opaque.
      required: false
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      description: Page size (1–100). Defaults to 50.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UUID:
      type: string
      format: uuid
    DateTime:
      type: string
      format: date-time
...
    Reminder:
      type: object
      required: [id, title, fire_at, status, created_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        batch_id:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2048
        fire_at:
          $ref: '#/components/schemas/DateTime'
        status:
          type: string
          enum: [scheduled, fired, completed, snoozed, dismissed, cancelled]
        fired_at:
          $ref: '#/components/schemas/DateTime'
        completed_at:
          $ref: '#/components/schemas/DateTime'
        suggested_event_kind:
          type: string
        created_at:
          $ref: '#/components/schemas/DateTime'
    Notification:
      type: object
      required: [id, kind, title, created_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        reminder_id:
          $ref: '#/components/schemas/UUID'
        kind:
          type: string
        title:
          type: string
        body:
          type: string
        url_path:
          type: string
        read_at:
          $ref: '#/components/schemas/DateTime'
        created_at:
          $ref: '#/components/schemas/DateTime'
    NotificationPrefs:
      type: object
      required: [push_enabled, apprise_enabled, apprise_url, email_enabled]
      properties:
        push_enabled:
          type: boolean
        apprise_enabled:
          type: boolean
          description: When true and `apprise_url` is set, fires notifications through the Apprise sidecar.
        apprise_url:
          type: string
          description: |
            Per-user Apprise URL (e.g. `mailto://`, `discord://`, `tgram://`).
            Empty string when unset. Schemes that turn the Apprise sidecar
            into a generic webhook target (`http`, `https`, `json`, `jsons`,
            `xml`, `xmls`, `form`, `forms`) are rejected unless the operator
            sets `APPRISE_ALLOW_WEBHOOK_SCHEMES=true`. URLs whose host
            resolves into a default-blocked private/loopback/IMDS range are
            also rejected; the operator can punch holes via
            `APPRISE_ALLOWED_HOST_CIDRS`.
        email_enabled:
          type: boolean
          description: |
            When true, the instance sends reminder emails directly via the
            operator-configured SMTP relay (`SMTP_HOST` et al.). Destination
            is the account email. Independent of the Apprise channel — both
            can be enabled, or just one. No-op if the operator hasn't
            configured SMTP.
        quiet_hours_start:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: "HH:MM in the user's timezone (set under `/api/me/prefs`)"
        quiet_hours_end:
          type: string
          pattern: '^\d{2}:\d{2}$'
          description: "HH:MM in the user's timezone (set under `/api/me/prefs`)"
...
paths:
...
  /api/batches/{id}/reminders:
    post:
      tags: [batches]
      operationId: createReminder
      summary: Create a manual reminder for a batch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, fire_at]
              properties:
                title:
                  type: string
                  maxLength: 200
                description:
                  type: string
                  maxLength: 2048
                fire_at:
                  type: string
                  format: date-time
                  description: RFC3339 timestamp
                suggested_event_kind:
                  type: string
      responses:
        "201":
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Batch not found or not owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags: [batches]
      operationId: listReminders
      summary: List reminders for a batch, ordered by fire_at ASC
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reminder'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Batch not found or not owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/batches/{id}/reminders/{reminderId}:
    patch:
      tags: [batches]
      operationId: updateReminder
      summary: Update a reminder (COALESCE). Allowed statuses for manual transition — scheduled, snoozed, completed, dismissed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: reminderId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                description:
                  type: string
                  maxLength: 2048
                fire_at:
                  type: string
                  format: date-time
                status:
                  type: string
                  enum: [scheduled, snoozed, completed, dismissed]
                suggested_event_kind:
                  type: string
      responses:
        "200":
          description: Updated reminder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reminder'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Reminder not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [batches]
      operationId: deleteReminder
      summary: Cancel a reminder (sets status=cancelled). 404 if already fired or completed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: reminderId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Cancelled
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Reminder not found or already in terminal state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Notifications -----------------------------------------------------------
  /api/notifications:
    get:
      tags: [notifications]
      operationId: listNotifications
      summary: In-app notification inbox (newest first)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Notification page. `next_cursor` is null when there are no more pages.
          content:
            application/json:
              schema:
                type: object
                required: [notifications, next_cursor]
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  next_cursor:
                    type: string
                    nullable: true
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications/read-all:
    post:
      tags: [notifications]
      operationId: markAllNotificationsRead
      summary: Mark all unread notifications as read
      responses:
        "204":
          description: Done
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications/{id}/read:
    post:
      tags: [notifications]
      operationId: markNotificationRead
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Marked read
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications/prefs:
    get:
      tags: [notifications]
      operationId: getNotificationPrefs
      summary: Get notification preferences (defaults if not set)
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPrefs'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [notifications]
      operationId: updateNotificationPrefs
      summary: Upsert notification preferences (COALESCE)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                push_enabled:
                  type: boolean
                apprise_enabled:
                  type: boolean
                apprise_url:
                  type: string
                  description: Empty string clears the saved URL.
                email_enabled:
                  type: boolean
                quiet_hours_start:
                  type: string
                  pattern: '^\d{2}:\d{2}$'
                quiet_hours_end:
                  type: string
                  pattern: '^\d{2}:\d{2}$'
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPrefs'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "422":
          description: Invalid time format or apprise_url (e.g. disallowed scheme or host in a blocked range)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...